14 improve distribution options

+++ tests/test_commands.py
@@ -6,7 +6,7 @@ import random
 import re
 import json
 import shutil
-from sjournal import SJournal, Note
+from src.sjournal import SJournal
 from utils_test import backup_file, delete_file, \
     get_project_root, \
     validate_note, validate_config, \
@@ -15,8 +15,7 @@ from logger import logger
 
 # Suite of testing to validate the CLI interface for sjournal using subprocess to call the application
 
-ROOT_DIR = get_project_root()
-sjournal_py = f"{os.path.join(ROOT_DIR, 'sjournal.py')}"
+
 
 n_gen_notes = 21
 n_gen_categories = 3
@@ -24,15 +23,25 @@ n_gen_styles = 5
 
 
 @pytest.fixture(scope="function")
-def clean_journal():
+def clean_journal(environment):
+    ROOT_DIR, HOME_DIR, SJOURNAL_DIR, DEBUG_OUTPUT, sjournal_exec = environment
 
-    logger.info(f"setup clean journal at {os.path.join(ROOT_DIR, 'journals', 'automated_test.db')}")
-    # backup current config file
-    config_filename = os.path.join(ROOT_DIR, "config.json")
-    backup_config_file = backup_file(config_filename)
+    # Make the reports directory in ~/sjournal/ if it does not exist
+    if not os.path.isdir(os.path.dirname(DEBUG_OUTPUT)):
+        os.makedirs(os.path.dirname(DEBUG_OUTPUT))
+
+    logger.info(f"setup clean journal at {os.path.join(HOME_DIR, 'sjournal', 'journals', 'automated_test.db')}")
+    # backup current config file if one exists
+    config_filename = os.path.join(SJOURNAL_DIR, "sjournal_config.json")
+    if os.path.isfile(config_filename):
+        backup_config_file = backup_file(config_filename)
+        config_file_existed = True
+    else:
+        backup_config_file = False
+        config_file_existed = False
 
     # if "automated_test.db" exists, delete it
-    journal_file = os.path.join(ROOT_DIR, "journals", "automated_test.db")
+    journal_file = os.path.join(SJOURNAL_DIR, "journals", "automated_test.db")
     if os.path.isfile(journal_file):
         delete_file(journal_file)
 
@@ -44,18 +53,24 @@ def clean_journal():
     yield journal
 
     # delete notebook
-    journal_file = os.path.join(ROOT_DIR, "journals", "automated_test.db")
+    journal_file = os.path.join(SJOURNAL_DIR, "journals", "automated_test.db")
     delete_file(journal_file)
 
-    with open(config_filename, "r") as config_file:
-        config = json.load(config_file)
-
-    # restore config file
-    backup_file(backup_config_file, config_filename)
-    delete_file(backup_config_file)
+    # restore default config file
+    if config_file_existed:
+        backup_file(backup_config_file, config_filename)
+        delete_file(backup_config_file)
+    else:
+        config = {
+            "journal_dir": os.path.join(HOME_DIR, "sjournal", "journals"),
+            "journal_name": "notes"
+        }
+        confstring = json.dumps(config)
+        with open(config_filename, "w") as config_file:
+            config_file.write(confstring)
 
     # delete any backups used for test
-    backup_dir = os.path.join(ROOT_DIR, "journals", "backups", "automated_test")
+    backup_dir = os.path.join(SJOURNAL_DIR, "journals", "backups", "automated_test")
     if os.path.isdir(backup_dir):
         shutil.rmtree(backup_dir)
 
@@ -104,7 +119,9 @@ def random_journal(clean_journal):
     yield journal
 
 
-def test_load(clean_journal):
+def test_load(clean_journal, environment):
+    ROOT_DIR, HOME_DIR, SJOURNAL_DIR, DEBUG_OUTPUT, sjournal_exec = environment
+
     # Start with clean empty journal
     journal = clean_journal
     logger.info("Clean journal should have the correct name")
@@ -115,21 +132,21 @@ def test_load(clean_journal):
     assert journal.length == 0
 
     # Verify that the correct config info is used with starting journal
-    validate_config({"journal_dir": "journals", "journal_name": "automated_test"})
+    validate_config({"journal_dir": os.path.join(SJOURNAL_DIR, "journals"), "journal_name": "automated_test"})
 
     # Load a new journal
     logger.info("Load a new journal named 'delete_this_journal'")
-    commandline = f'python {sjournal_py} load delete_this_journal'
+    commandline = f'{sjournal_exec} load delete_this_journal'
     result = subprocess.run(commandline, capture_output=True)
     logger.debug(f"stdout: {result.stdout}")
     assert result.returncode == 0
 
     # Verify that the correct config info is used with new journal
-    validate_config({"journal_dir": "journals", "journal_name": "delete_this_journal"})
+    validate_config({"journal_dir": os.path.join(SJOURNAL_DIR, "journals"), "journal_name": "delete_this_journal"})
 
     # Delete the new journal
     logger.info("Delete the new journal file")
-    new_journal_file = os.path.join(ROOT_DIR, "journals", "delete_this_journal.db")
+    new_journal_file = os.path.join(SJOURNAL_DIR, "journals", "delete_this_journal.db")
     delete_file(new_journal_file)
 
 
@@ -150,15 +167,15 @@ def test_load(clean_journal):
         (f'add -s "bold red" "Hello World"',
          {"category":"General", "content":"[bold red]Hello World[/]", "id":0}),
 ])
-def test_add_note(clean_journal, command, expected):
+def test_add_note(clean_journal, environment, command, expected):
+    ROOT_DIR, HOME_DIR, SJOURNAL_DIR, DEBUG_OUTPUT, sjournal_exec = environment
     # Start with clean empty journal
     journal = clean_journal
-
     for i in range(5):
         # Add note via commandline
         logger.info(f"Add note {i} via commandline")
 
-        commandline = f"python {sjournal_py} " + command
+        commandline = f"{sjournal_exec} " + command
 
         logger.debug(f"commandline: {commandline}")
         result = subprocess.run(commandline, capture_output=False)
@@ -174,7 +191,12 @@ def test_add_note(clean_journal, command, expected):
         validate_note(journal.notes[-1], expected_i)
 
 
-def test_edit_note(fixed_notes_journal):
+def test_edit_note(fixed_notes_journal, environment):
+    ROOT_DIR, HOME_DIR, SJOURNAL_DIR, DEBUG_OUTPUT, sjournal_exec = environment
+    for e in environment:
+        logger.info(f"{e=}")
+
+
     # Start with a journal that contains a few notes
     journal = fixed_notes_journal
     notes = journal.notes
@@ -184,8 +206,9 @@ def test_edit_note(fixed_notes_journal):
 
         # Execute Command with subsequent input
         logger.info(f"EDIT NOTE {note.id}")
+        logger.info([sjournal_exec.split()[0], sjournal_exec.split()[1], 'edit', f'{note.id}'.strip()])
 
-        proc = subprocess.Popen(['python', sjournal_py, 'edit', f'{note.id}'.strip()], stdout=subprocess.PIPE, stdin=subprocess.PIPE, stderr=subprocess.PIPE)
+        proc = subprocess.Popen([sjournal_exec.split()[0], sjournal_exec.split()[1], 'edit', f'{note.id}'.strip()], stdout=subprocess.PIPE, stdin=subprocess.PIPE, stderr=subprocess.PIPE)
         result = proc.communicate(input=b'EDITED')
 
         logger.debug(result)
@@ -198,7 +221,9 @@ def test_edit_note(fixed_notes_journal):
         validate_note(journal.notes[note.id], expected)
 
 
-def test_delete_note(fixed_notes_journal):
+def test_delete_note(fixed_notes_journal, environment):
+    ROOT_DIR, HOME_DIR, SJOURNAL_DIR, DEBUG_OUTPUT, sjournal_exec = environment
+
     # Start with a journal that contains a few notes
     journal = fixed_notes_journal
     original_notes = journal.notes
@@ -212,7 +237,7 @@ def test_delete_note(fixed_notes_journal):
         logger.info(f"Delete note #{note.id}")
 
         # Execute Command
-        commandline = f'python {sjournal_py} delete {note.id}'
+        commandline = f'{sjournal_exec} delete {note.id}'
         result = subprocess.run(commandline, capture_output=True)
 
         logger.debug(f"stdout: {result.stdout}")
@@ -247,12 +272,13 @@ def test_delete_note(fixed_notes_journal):
         'list -a -r -c',
         'list 10 -a -r -c',
 ])
-def test_list_default(random_journal, command):
+def test_list_default(random_journal, environment, command):
+    ROOT_DIR, HOME_DIR, SJOURNAL_DIR, DEBUG_OUTPUT, sjournal_exec = environment
 
     journal = random_journal
     notes = journal.notes
 
-    commandline = f"python {sjournal_py} --debug " + command
+    commandline = f"{sjournal_exec} --debug " + command
 
     n_printed = 5
     if '-c' in command:
@@ -277,8 +303,7 @@ def test_list_default(random_journal, command):
     assert result.returncode == 0
 
     # Read debug text into string
-    debug_file = os.path.join(ROOT_DIR, "reports", "debug.log")
-    with open(debug_file, "r") as output_file:
+    with open(DEBUG_OUTPUT, "r") as output_file:
         full_text = output_file.read()
 
     logger.debug(full_text)
@@ -301,13 +326,15 @@ def test_list_default(random_journal, command):
         'categories',
         'categories -s',
 ])
-def test_categories(random_journal, command):
+def test_categories(random_journal, environment, command):
+    ROOT_DIR, HOME_DIR, SJOURNAL_DIR, DEBUG_OUTPUT, sjournal_exec = environment
+
     journal = random_journal
     notes = journal.notes
     expected_categories = list(set([note.category for note in notes]))
 
     # List categories via command line
-    commandline = f"python {sjournal_py} --debug " + command
+    commandline = f"{sjournal_exec} --debug " + command
     if "-s" in command:
         commandline += f' "{expected_categories[0]}"'
         expected_categories = [expected_categories[0]]
@@ -318,8 +345,7 @@ def test_categories(random_journal, command):
     assert result.returncode == 0
 
     # Read debug text into string
-    debug_file = os.path.join(ROOT_DIR, "reports", "debug.log")
-    with open(debug_file, "r") as output_file:
+    with open(DEBUG_OUTPUT, "r") as output_file:
         full_text = output_file.read()
 
     logger.debug(full_text)
@@ -336,13 +362,15 @@ def test_categories(random_journal, command):
         'backup',
         'backup -f delete_this_backup',
 ])
-def test_backup(random_journal, command):
+def test_backup(random_journal, environment, command):
+    ROOT_DIR, HOME_DIR, SJOURNAL_DIR, DEBUG_OUTPUT, sjournal_exec = environment
+
     # Start with populated journal
     original_journal = random_journal
     original_notes = original_journal.notes
 
     # Backup journal via command line
-    commandline = f"python {sjournal_py} " + command
+    commandline = f"{sjournal_exec} " + command
     if "-f" in command:
         expected_backup_filename = command.split()[-1]
     else:
@@ -354,7 +382,7 @@ def test_backup(random_journal, command):
     assert result.returncode == 0
 
     # Verify that the backup exists
-    backup_dir = os.path.join(ROOT_DIR, "journals", "backups", "automated_test")
+    backup_dir = os.path.join(SJOURNAL_DIR, "journals", "backups", "automated_test")
     assert os.path.isdir(backup_dir)
 
     with os.scandir(backup_dir) as dir_contents:
@@ -371,7 +399,7 @@ def test_backup(random_journal, command):
 
     # Load the backup and confirm that all original notes are present
     # Change config file to reflect backup location
-    config_file_path = os.path.join(ROOT_DIR, "config.json")
+    config_file_path = os.path.join(SJOURNAL_DIR, "sjournal_config.json")
     config = {
         "journal_dir": os.path.join("journals", "backups", "automated_test"),
         "journal_name": backup_journal_filename
@@ -398,7 +426,9 @@ def test_backup(random_journal, command):
         # ('restore', 'delete -j'),
         # ('restore -f delete_this_backup', 'delete -j'),
 ])
-def test_restore(random_journal, command, action):
+def test_restore(random_journal, environment, command, action):
+    ROOT_DIR, HOME_DIR, SJOURNAL_DIR, DEBUG_OUTPUT, sjournal_exec = environment
+
     # Start with populated journal
     original_journal = random_journal
     original_notes = original_journal.notes
@@ -430,7 +460,7 @@ def test_restore(random_journal, command, action):
     assert original_journal.length == 0
 
     # Restore the journal
-    commandline = f"python {sjournal_py} " + command
+    commandline = f"{sjournal_exec} " + command
     logger.debug(commandline)
     result = subprocess.run(commandline, capture_output=False)
     logger.debug(result)
@@ -442,12 +472,14 @@ def test_restore(random_journal, command, action):
         assert note == original_notes[i], f"New Note does not equal old Note:\nNew: {note}\nOld: {original_notes[i]}"
 
 
-def test_erase(random_journal):
+def test_erase(random_journal, environment):
+    ROOT_DIR, HOME_DIR, SJOURNAL_DIR, DEBUG_OUTPUT, sjournal_exec = environment
+
     journal = random_journal
     assert journal.length == n_gen_notes
 
     # Erase the notebook via the command line
-    commandline = f"python {sjournal_py} --debug erase"
+    commandline = f"{sjournal_exec} --debug erase"
     logger.debug(commandline)
     result = subprocess.run(commandline, capture_output=False)
     logger.debug(result)
@@ -459,26 +491,27 @@ def test_erase(random_journal):
         'search Note 1',
         'search',
 ])
-def test_search(fixed_notes_journal, command):
+def test_search(fixed_notes_journal, environment, command):
+    ROOT_DIR, HOME_DIR, SJOURNAL_DIR, DEBUG_OUTPUT, sjournal_exec = environment
+
     # Start with populated, non-random journal
     journal = fixed_notes_journal
     notes = journal.notes
     for note in notes:
         logger.debug(note)
     # Search the notes via commandline with debug
-    commandline = f"python {sjournal_py} --debug " + command
+    commandline = f"{sjournal_exec} --debug " + command
 
     if "Note" not in command:
         for i, note in enumerate(notes):
             # send command
-            commandline = f"python {sjournal_py} --debug {command} {i}"
+            commandline = f"{sjournal_exec} --debug {command} {i}"
             result = subprocess.run(commandline, capture_output=False)
             logger.debug(result)
             assert result.returncode == 0
 
             # Read debug text into string
-            debug_file = os.path.join(ROOT_DIR, "reports", "debug.log")
-            with open(debug_file, "r") as output_file:
+            with open(DEBUG_OUTPUT, "r") as output_file:
                 full_text = output_file.read()
 
             logger.debug(full_text)
@@ -496,8 +529,7 @@ def test_search(fixed_notes_journal, command):
         assert result.returncode == 0
 
         # Read debug text into string
-        debug_file = os.path.join(ROOT_DIR, "reports", "debug.log")
-        with open(debug_file, "r") as output_file:
+        with open(DEBUG_OUTPUT, "r") as output_file:
             full_text = output_file.read()
 
         logger.debug(full_text)
@@ -510,3 +542,13 @@ def test_search(fixed_notes_journal, command):
             match = re.search(regex, full_text)
             assert match, f"Could not find Note # {note.id} in debug output\n{regex}"
             logger.debug(match.group(0))
+
+
+def test_dummy(environment):
+    ROOT_DIR, HOME_DIR, SJOURNAL_DIR, DEBUG_OUTPUT, sjournal_exec = environment
+
+    logger.info(f"{ROOT_DIR=}")
+    logger.info(f"{HOME_DIR=}")
+    logger.info(f"{SJOURNAL_DIR=}")
+    logger.info(f"{DEBUG_OUTPUT=}")
+    logger.info(f"{sjournal_exec=}")
